requirements.txt
absl-py==1.4.0
anyio==3.6.2
astunparse==1.6.3
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
cryptography @ file:///croot/cryptography_1673298753778/work
fastapi==0.95.0
FastAPI-SQLAlchemy==0.2.1
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
idna @ file:///croot/idna_1666125576474/work
jax==0.4.8
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
protobuf==4.22.3
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
pytz==2023.3
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
toolz @ file:///croot/toolz_1667464077321/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
typing_extensions==4.5.0
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1673575502006/work
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.14.1
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
